Move dumper to not be included in star import
